// client/src/generated/models/ErrorResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * envlink
 * A fast and reliable URL shortening service built with NestJS, enabling easy link generation, management, and usage tracking.
 *
 * The version of the OpenAPI document: v1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface ErrorResponse
 */
export interface ErrorResponse {
  /**
   *
   * @type {boolean}
   * @memberof ErrorResponse
   */
  success?: boolean;
  /**
   *
   * @type {number}
   * @memberof ErrorResponse
   */
  status?: number;
  /**
   *
   * @type {string}
   * @memberof ErrorResponse
   */
  message?: string;
  /**
   *
   * @type {string}
   * @memberof ErrorResponse
   */
  error?: string | null;
  /**
   *
   * @type {string}
   * @memberof ErrorResponse
   */
  path: string;
  /**
   *
   * @type {Date}
   * @memberof ErrorResponse
   */
  timestamp: Date;
}

/**
 * Check if a given object implements the ErrorResponse interface.
 */
export function instanceOfErrorResponse(value: object): value is ErrorResponse {
  if (!("path" in value) || value["path"] === undefined) return false;
  if (!("timestamp" in value) || value["timestamp"] === undefined) return false;
  return true;
}

export function ErrorResponseFromJSON(json: any): ErrorResponse {
  return ErrorResponseFromJSONTyped(json, false);
}

export function ErrorResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): ErrorResponse {
  if (json == null) {
    return json;
  }
  return {
    success: json["success"] == null ? undefined : json["success"],
    status: json["status"] == null ? undefined : json["status"],
    message: json["message"] == null ? undefined : json["message"],
    error: json["error"] == null ? undefined : json["error"],
    path: json["path"],
    timestamp: new Date(json["timestamp"]),
  };
}

export function ErrorResponseToJSON(json: any): ErrorResponse {
  return ErrorResponseToJSONTyped(json, false);
}

export function ErrorResponseToJSONTyped(
  value?: ErrorResponse | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    success: value["success"],
    status: value["status"],
    message: value["message"],
    error: value["error"],
    path: value["path"],
    timestamp: value["timestamp"].toISOString(),
  };
}